requirements.txt
streamlit
plotly
google-generativeai
newsapi-python
beautifulsoup4
requests
python-dotenv
pandas
numpy
lxml
textblob
scikit-learn
joblib
